r#78. Rename package oscal to package generate for consistency
