+++ src/internal/generate/generate_ref_queue.go
@@ -1,4 +1,4 @@
-package oscal
+package generate
 
 type RefQueue struct {
 	refs       []string
