+++ src/internal/generate/generate_test.go
@@ -1,4 +1,4 @@
-package oscal
+package generate
 
 import (
 	"go/format"
